requirements.txt
azure-functions==1.18.0
azure-functions-durable==1.2.9
azure-identity>=1.15.0
